+++ term_image/image/kitty.py
@@ -16,8 +16,10 @@ from ..exceptions import _style_error
 from ..utils import CSI, ESC, ST, lock_tty, query_terminal
 from .common import GraphicsImage
 
-FORMAT_SPEC = re.compile(r"([^zm]*)(z(-?\d+)?)?(m[01])?(.*)", re.ASCII)
-# Constants for ``KittyImage`` render method
+FORMAT_SPEC = re.compile(
+    r"([^LWzmc]*)([LW])?(z(-?\d+)?)?(m[01])?(c[0-9])?(.*)", re.ASCII
+)
+# Constants for render methods
 LINES = "lines"
 WHOLE = "whole"
 
@@ -61,7 +63,16 @@ class KittyImage(GraphicsImage):
 
     ::
 
-        [ z [index] ] [ m {0 | 1} ]
+        [method] [ z [index] ] [ m {0 | 1} ] [ c {0-9} ]
+
+    * ``method``: Render method override.
+
+      Can be one of:
+
+        * ``L``: **LINES** render method (current frame only, for animated images).
+        * ``W``: **WHOLE** render method (current frame only, for animated images).
+
+      Default: Current effective render method of the image.
 
     * ``z``: Image/Text stacking order.
 
@@ -72,8 +83,8 @@ class KittyImage(GraphicsImage):
 
         * ``>= 0``, the image will be drawn above text.
         * ``< 0``, the image will be drawn below text.
-        * ``< -(2 ** 31) / 2``, the image will be drawn below non-default text
-          background colors.
+        * ``< -(2**31)/2``, the image will be drawn below cells with non-default
+          background color.
 
       * ``z`` without ``index`` is currently only used internally.
       * If *absent*, defaults to z-index ``0``.
@@ -90,6 +101,12 @@ class KittyImage(GraphicsImage):
       * If *absent*, defaults to ``m0``.
       * e.g ``m0``, ``m1``.
 
+    * ``c``: ZLIB compression level.
+
+      * 1 gives best speed, 9 gives best compression, 0 gives no compression at all.
+      * This results in a trade-off between render time and transmission size/time.
+      * If *absent*, defaults to ``c4``.
+      * e.g ``c0``, ``c9``.
 
     ATTENTION:
         Currently supported terminal emulators include:
@@ -102,6 +119,16 @@ class KittyImage(GraphicsImage):
     _default_render_method: str = LINES
     _render_method: str = LINES
     _style_args = {
+        "method": (
+            (
+                lambda x: isinstance(x, str),
+                "Render method must be a string",
+            ),
+            (
+                lambda x: x in KittyImage._render_methods,
+                "Unknown render method",
+            ),
+        ),
         "z_index": (
             (
                 lambda x: x is None or isinstance(x, int),
@@ -119,6 +146,16 @@ class KittyImage(GraphicsImage):
             ),
             (lambda _: True, ""),
         ),
+        "compress": (
+            (
+                lambda x: isinstance(x, int),
+                "Compression level must be an integer",
+            ),
+            (
+                lambda x: 0 <= x <= 9,
+                "Compression level must be between 0 and 9 (both inclusive)",
+            ),
+        ),
     }
 
     _KITTY_VERSION: Tuple[int, int, int] = ()
@@ -126,7 +163,12 @@ class KittyImage(GraphicsImage):
 
     # Only defined for the purpose of proper self-documentation
     def draw(
-        self, *args, z_index: Optional[int] = 0, mix: bool = False, **kwargs
+        self,
+        *args,
+        z_index: Optional[int] = 0,
+        mix: bool = False,
+        compress: int = 4,
+        **kwargs,
     ) -> None:
         """Draws an image to standard output.
 
@@ -141,8 +183,8 @@ class KittyImage(GraphicsImage):
 
               * ``>= 0``, the image will be drawn above text.
               * ``< 0``, the image will be drawn below text.
-              * ``< -(2 ** 31) / 2``, the image will be drawn below non-default text
-                background colors.
+              * ``< -(2**31)/2``, the image will be drawn below cells with
+                non-default background color.
               * ``None``, deletes any directly overlapping image.
 
               .. note::
@@ -158,6 +200,12 @@ class KittyImage(GraphicsImage):
               * ``False``, text within the region covered by the image will be
                 erased, though text can be inter-mixed with the image after it's
                 been drawn.
+
+            compress: ZLIB compression level.
+
+              An integer between 0 and 9: 1 gives best speed, 9 gives best compression,
+              0 gives no compression at all. This results in a trade-off between render
+              time and transmission size/time.
 
             kwargs: Keyword arguments passed up the inheritance chain.
 
@@ -225,7 +273,9 @@ class KittyImage(GraphicsImage):
 
     @classmethod
     def _check_style_format_spec(cls, spec: str, original: str) -> Dict[str, Any]:
-        parent, z, index, mix, invalid = FORMAT_SPEC.fullmatch(spec).groups()
+        parent, method, z, index, mix, compress, invalid = FORMAT_SPEC.fullmatch(
+            spec
+        ).groups()
         if invalid:
             raise _style_error(cls)(
                 f"Invalid style-specific format specification {original!r}"
@@ -234,10 +284,14 @@ class KittyImage(GraphicsImage):
         args = {}
         if parent:
             args.update(super()._check_style_format_spec(parent, original))
+        if method:
+            args["method"] = LINES if method == "L" else WHOLE
         if z:
             args["z_index"] = index and int(index)
         if mix:
             args["mix"] = bool(int(mix[-1]))
+        if compress:
+            args["compress"] = int(compress[-1])
 
         return cls._check_style_args(args)
 
@@ -289,8 +343,10 @@ class KittyImage(GraphicsImage):
         alpha: Union[None, float, str],
         *,
         frame: bool = False,
+        method: Optional[str] = None,
         z_index: Optional[int] = 0,
         mix: bool = False,
+        compress: int = 4,
     ) -> str:
         # NOTE: It's more efficient to write separate strings to the buffer separately
         # than concatenate and write together.
@@ -300,6 +356,7 @@ class KittyImage(GraphicsImage):
         # Since we use `c` and `r` control data keys, there's no need upscaling the
         # image on this end; ensures minimal payload.
 
+        render_method = method or self._render_method
         r_width, r_height = self.rendered_size
         width, height = self._get_minimal_render_size()
 
@@ -320,23 +377,29 @@ class KittyImage(GraphicsImage):
             delete = f"{START}a=d,d=c;{ST}"
             clear = f"{delete}{ESC}7{CSI}{r_width}C{delete}{ESC}8"
 
-        if self._render_method == LINES:
+        if render_method == LINES:
             cell_height = height // r_height
             bytes_per_line = width * cell_height * (format // 8)
             vars(control_data).update(dict(v=cell_height, r=1))
 
             with io.StringIO() as buffer, io.BytesIO(raw_image) as raw_image:
-                trans = Transmission(control_data, raw_image.read(bytes_per_line))
+                trans = Transmission(
+                    control_data, raw_image.read(bytes_per_line), compress
+                )
                 z_index is None and buffer.write(clear)
-                buffer.write(trans.get_chunked())
+                for chunk in trans.get_chunks():
+                    buffer.write(chunk)
                 # Writing spaces clears any text under transparent areas of an image
                 for _ in range(r_height - 1):
                     buffer.write(erase)
                     buffer.write(jump_right)
                     buffer.write("\n")
-                    trans = Transmission(control_data, raw_image.read(bytes_per_line))
+                    trans = Transmission(
+                        control_data, raw_image.read(bytes_per_line), compress
+                    )
                     z_index is None and buffer.write(clear)
-                    buffer.write(trans.get_chunked())
+                    for chunk in trans.get_chunks():
+                        buffer.write(chunk)
                 buffer.write(erase)
                 buffer.write(jump_right)
 
@@ -346,7 +409,7 @@ class KittyImage(GraphicsImage):
             return "".join(
                 (
                     z_index is None and clear or "",
-                    Transmission(control_data, raw_image).get_chunked(),
+                    Transmission(control_data, raw_image, compress).get_chunked(),
                     f"{erase}{jump_right}\n" * (r_height - 1),
                     f"{erase}{jump_right}",
                 )
@@ -360,19 +423,27 @@ class Transmission:
     Args:
         control: The control data.
         payload: The payload.
+        level: Compression level.
     """
 
     control: ControlData
     payload: bytes
 
+    # From tests with a few images, from anything beyond 4, the decrease in size
+    # doesn't seem to be worth the increase in compression time in most cases.
+    # Might change if proven otherwise.
+    level: int = 4
+
     def __post_init__(self):
         self._compressed = False
-        if self.control.o == o.ZLIB:
+        if self.level:
             self.compress()
+        else:
+            self.control.o = None
 
     def compress(self):
-        if self.control.t == t.DIRECT and not self._compressed:
-            self.payload = compress(self.payload)
+        if self.control.t == t.DIRECT and not self._compressed and self.level:
+            self.payload = compress(self.payload, self.level)
             self.control.o = o.ZLIB
             self._compressed = True
 
@@ -464,7 +535,7 @@ class ControlData:
     s: Optional[int] = None  # image width
     v: Optional[int] = None  # image height
     z: Optional[int] = z.IN_FRONT  # z-index
-    o: Optional[str] = o.ZLIB  # compression
+    o: Optional[str] = None  # compression
     C: Optional[int] = C.STAY  # cursor movement policy
 
     # # Image display size in columns and rows/lines